Move spotify into datasources folder
