+++ src/app.js
@@ -4,7 +4,7 @@ const logger = require('morgan');
 const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 const cors = require('cors');
-const SpotifyAPI = require('./spotify');
+const SpotifyAPI = require('./datasources/spotify/SpotifyAPI');
 const routes = require('./routes/index');
 const {typeDefs, resolvers} = require('./schema');
 const {haveToken} = require('./schema/resolvers')
